Add tranche edit section
