+++ centrifuge-js/src/modules/pools.ts
@@ -661,7 +661,7 @@ export type PoolReport = {
   }
   uri: string
 }
-interface TrancheFormValues {
+export interface TrancheFormValues {
   tokenName: string
   symbolName: string
   interestRate: number | ''
